+++ mod/Mods/tech/DATIS/Options/optionsDb.lua
@@ -1,5 +1,6 @@
-local DbOption = require('Options.DbOption')
+local DbOption = require("Options.DbOption")
 
 return {
-    gcloudAccessKey = DbOption.new():setValue(''):editbox(),
+    gcloudAccessKey = DbOption.new():setValue(""):editbox(),
+    debugLoggingEnabled = DbOption.new():setValue(false):checkbox()
 }