+++ CHANGELOG.md
@@ -5,12 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Option to enable debug logging (useful when investigating issues; logs into `Saved Games\DCS\Logs\DATIS.log`)
+
 ### Fixed
 - Fix rounding issue in parsing of certain frequencies #15
 
 ## [0.8.0] - 2019-04-23
 ### Changed
-- DATIS now acts as an SRS version 1.6.0.0 client (and is thus not backwards compatible to <1.6 servers anymore)
+- DATIS now acts as an SRS version 1.6.0.0 client (and thus doesn't work with <1.6 servers anymore)
 
 ## [0.7.1] - 2019-03-03
 ### Fixed